main

+++ Birthday-Reminder.py
@@ -4,57 +4,51 @@ from pystyle import *
 
 from CSV_by_Laix import *
 
+
 print("\033[91m" + "[*] Installing missing modules" + "\033[97m") 
 
-if platform.system().startswith("Windows"):
-    try:
-        import csv
-    except ImportError:
-        os.system("py -m pip install csv -q -q -q")
-        import csv
+try:
+    import csv
+except ImportError:
+    os.system("py -m pip install csv -q -q -q")
+    import csv
 
-    try:
-        import ctypes
-    except ImportError:
-        os.system("py -m pip install ctypes -q -q -q")
-        import ctypes
+try:
+    import ctypes
+except ImportError:
+    os.system("py -m pip install ctypes -q -q -q")
+    import ctypes
 
-    try:
-        import datetime
-    except ImportError:
-        os.system("py -m pip install datetime -q -q -q")
-        import datetime
+try:
+    import datetime
+except ImportError:
+    os.system("py -m pip install datetime -q -q -q")
+    import datetime
 
-    try:
-        import winshell
-    except ImportError:
-        os.system("py -m pip install winshell -q -q -q")
-        os.system("py -m pip install pywin32 -q -q -q")
-        import winshell
+try:
+    import winshell
+except ImportError:
+    os.system("py -m pip install winshell -q -q -q")
+    os.system("py -m pip install pywin32 -q -q -q")
+    import winshell
 
-    try:
-        import logging
-    except ImportError:
-        os.system("py -m pip install logging -q -q -q")
-        import logging
+try:
+    import logging
+except ImportError:
+    os.system("py -m pip install logging -q -q -q")
+    import logging
 
-    try:
-        import colorama
-    except ImportError:
-        os.system("py -m pip install colorama -q -q -q")
-        from colorama import Fore, Back, Style
+try:
+    import colorama
+except ImportError:
+    os.system("py -m pip install colorama -q -q -q")
+    from colorama import Fore, Back, Style
 
-    try:
-        import pystyle
-    except ImportError:
-        os.system("py -m pip install pystyle -q -q -q")
-        from pystyle import *
-
-elif platform.system().startswith("Linux"):
-    print("[*] The Linux version is not yet available. You must run the program on Windows or MacOS.")
-
-elif platform.system().startswith("Darwin"):
-    print("[*] The MacOS version is not yet available. You must run the program on Windows or MacOS.")
+try:
+    import pystyle
+except ImportError:
+    os.system("py -m pip install pystyle -q -q -q")
+    from pystyle import *
 
 
 logging.basicConfig(
@@ -111,50 +105,57 @@ def remove_person():
 
 
 def WStartup(src_file):
-    startup_folder = winshell.startup()
-    shortcut_path = os.path.join(startup_folder, f'{src_file}.lnk')
-    script_directory = os.path.dirname(os.path.abspath(__file__))
+    if platform.system().startswith("Windows"):
+
+        startup_folder = winshell.startup()
+        shortcut_path = os.path.join(startup_folder, f'{src_file}.lnk')
+        script_directory = os.path.dirname(os.path.abspath(__file__))
 
 
-    if os.path.exists(shortcut_path):
-        print("[*] Windows Startup option is \033[92menabled\033[97m, do you want to \033[91mdisable\033[97m it?")
-        print('\033[91mY\033[97m or \033[92mN\033[97m')
-        YorN = input("> ")
+        if os.path.exists(shortcut_path):
+            print("[*] Windows Startup option is \033[92menabled\033[97m, do you want to \033[91mdisable\033[97m it?")
+            print('\033[91mY\033[97m or \033[92mN\033[97m')
+            YorN = input("> ")
 
-        if YorN.lower() == 'y':
-            os.remove(shortcut_path)
-            print("[*] Windows Startup option is now \033[91mdisabled.\033[97m")
-            logging.info("Windows Startup option is now disabled.")
+            if YorN.lower() == 'y':
+                os.remove(shortcut_path)
+                print("[*] Windows Startup option is now \033[91mdisabled.\033[97m")
+                logging.info("Windows Startup option is now disabled.")
 
-        elif YorN.lower() == 'n':
-            pass
+            elif YorN.lower() == 'n':
+                pass
 
-        else:
-            print("\033[91m" + "Invalid choice. Please select a valid option." + "\033[97m")
-
-    else:
-        print("[*] Windows Startup option is \033[91mdisabled\033[97m, do you want to \033[92menable\033[97m it?")
-        print('\033[92mY\033[97m or \033[91mN\033[97m')
-        YorN = input("> ")
-
-        if YorN.lower() == 'y':
-            
-            winshell.CreateShortcut(
-                Path=shortcut_path, 
-                Target=os.path.abspath(src_file),
-                StartIn=script_directory,
-                Description="Script for birthday notification at startup")
-            
-
-            print("[*] Windows Startup option is now \033[92menabled.\033[97m")
-            logging.info("Windows Startup option is now enabled.")
-
-        elif YorN.lower() == 'n':
-            pass
+            else:
+                print("\033[91m" + "Invalid choice. Please select a valid option." + "\033[97m")
 
         else:
-            print("\033[91m" + "Invalid choice. Please select a valid option." + "\033[97m")
+            print("[*] Windows Startup option is \033[91mdisabled\033[97m, do you want to \033[92menable\033[97m it?")
+            print('\033[92mY\033[97m or \033[91mN\033[97m')
+            YorN = input("> ")
 
+            if YorN.lower() == 'y':
+                
+                winshell.CreateShortcut(
+                    Path=shortcut_path, 
+                    Target=os.path.abspath(src_file),
+                    StartIn=script_directory,
+                    Description="Script for birthday notification at startup")
+                
+
+                print("[*] Windows Startup option is now \033[92menabled.\033[97m")
+                logging.info("Windows Startup option is now enabled.")
+
+            elif YorN.lower() == 'n':
+                pass
+
+            else:
+                print("\033[91m" + "Invalid choice. Please select a valid option." + "\033[97m")
+    
+    elif platform.system().startswith("Linux"):
+        print("not avaible on linux")
+
+    elif platform.system().startswith("Darwin"):
+        print("not avaible on MacOs")
 
 
 def main():
